# src/main/webapp/docs/api.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: "GoShopping API"
  description: "This is a api"
servers:
  - url: 'http://localhost:4321/api'
paths:
...
  /item:
    get:
      security:
        - bearerAuth: [ ]
      summary: 'get/search items'
      parameters:
        - in: query
          name: text
          schema:
            type: string
          description: 'filter items with name containing this text'
        - in: query
          name: tag
          schema:
            type: string
          description: 'filter items with a tag equal to this'
        - in: query
          name: quantity
          schema:
            type: number
          description: 'how many items you want in the result set'
        - in: query
          name: page
          schema:
            type: number
          description: 'Without paging, you get a result set R. Now add this page parameter and the first item in the result set will be the (page * quantity)th item in R'
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemList"
    post:
      security:
        - bearerAuth: [ ]
      summary: create an item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostableItem"
      responses:
        200:
          description: "the created item"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /item/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: represents the item you want to do an HTTP method on
    get:
      security:
        - bearerAuth: [ ]
      summary: get an item
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    put:
      security:
        - bearerAuth: [ ]
      summary: update the item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutableItem"
      responses:
        200:
          description: "the updated item"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TagList:
      type: 'array'
      items:
        $ref: '#/components/schemas/Tag'
    Tag:
      type: 'object'
      required:
        - id
        - name
      properties:
        id:
          type: 'integer'
        name:
          type: "string"
    BareTagList:
      type: 'array'
      items:
        $ref: '#/components/schemas/BareTag'
    BareTag:
      type: 'object'
      required:
        - name
      properties:
        name:
          type: 'string'
    ItemList:
      type: 'array'
      items:
        $ref: '#/components/schemas/Item'
    Item:
      type: 'object'
      required:
        - id
        - name
        - price
        - description
        - tags
      properties:
        id:
          type: 'integer'
        name:
          type: "string"
        price:
          type: 'number'
        description:
          type: 'string'
        tags:
          $ref: '#/components/schemas/TagList'
    PostableItem:
      type: 'object'
      required:
        - name
        - price
        - description
        - tags
      properties:
        name:
          type: "string"
        price:
          type: 'number'
        description:
          type: 'string'
        tags:
          $ref: '#/components/schemas/BareTagList'
    PutableItem:
      type: 'object'
      required:
        - name
        - price
        - description
        - tags
      properties:
        name:
          type: "string"
        price:
          type: 'number'
        description:
          type: 'string'
        tags:
          $ref: '#/components/schemas/BareTagList'
...